ansible/roles/google_cloud/files/function/requirements.txt
google-cloud-storage
gsutil
affine==2.2.2
attrs==19.1.0
certifi==2019.6.16
cftime==1.0.3.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
geojson==2.4.1
netCDF4==1.5.1.2
numpy==1.16.4
pyparsing==2.4.0
pyugrid==0.3.1
rasterio==1.0.24
scipy==1.2.1
Shapely==1.6.4.post2
snuggs==1.4.6
tqdm==4.32.2
wincertstore==0.2
